fix: ピクチャ関連のファイルがない場合は処理を飛ばす

+++ scripts/deploy-demo.sh
@@ -38,7 +38,9 @@ unzip -d wwawing-dist wwawing-dist.zip
 
 DEST_DIR="./wwawing.com/unstable"
 cp "$SRC_DIR/mapdata/wwamap.dat" $DEST_DIR
-cp "$SRC_DIR/mapdata/picture_test.dat" $DEST_DIR
+if [ -e "$SRC_DIR/mapdata/picture_test.dat" ] ; then
+    cp "$SRC_DIR/mapdata/picture_test.dat" $DEST_DIR
+fi
 cp "$SRC_DIR/mapdata/wwa.js" $DEST_DIR
 cp "$SRC_DIR/mapdata/wwa.css" $DEST_DIR
 cp "$SRC_DIR/mapdata/style.css" $DEST_DIR
@@ -47,7 +49,9 @@ cp "$SRC_DIR/mapdata/cover.gif" $DEST_DIR
 cp -R "$SRC_DIR/mapdata/audio" $DEST_DIR
 
 # ピクチャ機能で使用
-cp "$SRC_DIR/mapdata/picture_test-picture.json" $DEST_DIR
-cp "$SRC_DIR/mapdata/wwawing-disp.png" $DEST_DIR
+if [ -e "$SRC_DIR/mapdata/picture_test-picture.json" ] ; then
+    cp "$SRC_DIR/mapdata/picture_test-picture.json" $DEST_DIR
+    cp "$SRC_DIR/mapdata/wwawing-disp.png" $DEST_DIR
+fi
 
 rm -rf $SRC_DIR ./wwawing-dist ./wwawing-dist.zip
